Actualizacion miercoles 11 de julio
Actualizar repositorio

+++ Model/mdlCupones.php
@@ -6,21 +6,12 @@ require_once('../scripts/Validaciones.php');
 
 class obtener_cupon {
 
-    //Se crea el metodo obtener los objetos revisados esto con el objetivo obtener aquellos cupones que han sido colocados por una empresa 
-    //en especificos
-    public function recibir_id() {
-        //Se recibe el dentificador de la empresa de  la clase contenido
-        $id_empresa = $_GET['ide'];
-        return $id_empresa;
-    }
-
     public function lista_cupones() {
         //Se llama a la case conectar del archivo conexion.php
         $this->dbh = new PDO('mysql:host=127.0.0.1:3306;dbname=campeche', "root", "P4SSW0RD");
         //Se declara la variable identificador el cual obtendra el id de la empresa el cual se obtiene del metodo lista_ojetos_revisados
-        $id_empresa = $_GET['ide'];
         //Se recibe l identificador de la empresa del metodo lista de objetos revisados
-        $sql = "select c.id_cupon, c.id_revision_objeto, c.titulo, c.descripcion_corta, c.descripcion_larga, c.id_imagen_extra, c.vigencia_inicio, c.terminos_y_condiciones, c.limite_codigos from (cupon c inner join revision_objeto on c.id_revision_objeto = revision_objeto.id_revision_objeto) inner join empresa on revision_objeto.id_empresa = '$id_empresa' group by titulo;";
+        $sql = "select c.id_cupon, c.id_revision_objeto, c.titulo, c.descripcion_corta, c.descripcion_larga, c.id_imagen_extra, c.vigencia_inicio, c.terminos_y_condiciones, c.limite_codigos from (cupon c inner join revision_objeto on c.id_revision_objeto = revision_objeto.id_revision_objeto) inner join empresa on revision_objeto.id_empresa = " . $_SESSION['idemp'] . " group by titulo;";
         foreach ($this->dbh->query($sql) as $res) {
             $this->platillo[] = $res;
         }
@@ -38,13 +29,11 @@ class obtener_cupon {
         $na = new validacion();
         $idro = $na->generar_aleatorio();
         $iie = $na->generar_alfanumerico();
-        //Se reciben los parametros de la vista addcupon
-        $id_empresa = $_POST['id_empresa'];
         //Fecha de creacion y hora 
         $fa = $na->fecha_actual();
         $status = 'C';
         $sql = "INSERT INTO revision_objeto(id_revision_objeto,id_empresa,fecha_creacion,status)
-        VALUES('$idro','$id_empresa','$fa','$status')";
+        VALUES('$idro'," . $_SESSION['idemp'] . ",'$fa','$status')";
         if (!mysqli_query($pd, $sql)) {
             die('Error: ' . mysqli_error($pd));
         }
@@ -59,7 +48,7 @@ class obtener_cupon {
         if (move_uploaded_file($_FILES['id_imagen_extra']['tmp_name'], "C:/xampp/htdocs/campeche-web2/Imagenes/Cupones/$nombreimagen")) {
             $msg = "El archivo ha sido cargado correctamente.<br>";
         } else {
-           $nombreimagen = ""; 
+            $nombreimagen = "";
         }
         $nombres = $nombreimagen;
         $vigencia_inicio = $_POST['vigencia_inicio'];
@@ -73,9 +62,8 @@ class obtener_cupon {
         if (!mysqli_query($pd, $sql2)) {
             die('Error: ' . mysqli_error($pd));
         }
-        $id = $id_empresa;
         mysqli_close($pd);
-        return $id;
+        header("Location:https://localhost/campeche-web2/Controller/ControladorSitios.php");
     }
 
     public function eliminar_cupon() {
@@ -84,11 +72,10 @@ class obtener_cupon {
         //se llama a la funcion con para obtener la variable conexion la cual sera utilizada para ejecutar la sentencia sql
         $pd = $conn->con();
         //Se obtienen los parametros de la vista del cupon
-        $id_empresa = $_GET["id_empresa"];
         $id_revision_objeto = $_GET["id_revision_objeto"];
         $id_cupon = $_GET["id_cupon"];
         $imagen = $_GET["id_imagen_extra"];
-        $Eliminar = "Delete from revision_objeto where id_revision_objeto = " . $id_revision_objeto . " AND id_empresa = " . $id_empresa . "";
+        $Eliminar = "Delete from revision_objeto where id_revision_objeto = " . $id_revision_objeto . " AND id_empresa = '" . $_SESSION['idemp'] . "'";
         $Eliminar2 = "Delete from cupon where id_cupon = " . $id_cupon . " and id_revision_objeto = " . $id_revision_objeto . "";
         if (!mysqli_query($pd, $Eliminar2)) {
             die('Error: ' . mysqli_error($pd));
@@ -96,11 +83,16 @@ class obtener_cupon {
         if (!mysqli_query($pd, $Eliminar)) {
             die('Error: ' . mysqli_error($pd));
         }
-        //Se elimina la imagen 
-        $ruta = "C:/xampp/htdocs/campeche-web2/Imagenes/Cupones/";
-        unlink($ruta . $imagen);
-        mysqli_close($pd);
-        header("Location:https://localhost/campeche-web2");
+        if ($imagen == "") {
+            mysqli_close($pd);
+            header("Location:https://localhost/campeche-web2/Controller/ControladorSitios.php");
+        } else {
+            //Se elimina la imagen 
+            $ruta = "C:/xampp/htdocs/campeche-web2/Imagenes/Cupones/";
+            unlink($ruta . $imagen);
+            mysqli_close($pd);
+            header("Location:https://localhost/campeche-web2/Controller/ControladorSitios.php");
+        }
     }
 
     public function actualizar_cupon() {
@@ -142,17 +134,12 @@ class obtener_cupon {
         $minuto_actual = date("i");
         $segundo_actual = date("s");
         $fecha_actual = $año_actual . "" . $mes_actual . "" . $dia_actual . "" . $hora_actual . "" . $minuto_actual . "" . $segundo_actual;
-        if (is_uploaded_file($uploadfile_temporal)) {
-            move_uploaded_file($uploadfile_temporal, $uploadfile_nombre);
-        } else {
-            echo "error";
-        }
         $vigencia_inicio = $_POST['vigencia_inicio'];
         $vigencia_fin = $_POST['vigencia_fin'];
         $terminos_y_condiciones = $_POST['terminos_y_condiciones'];
         $nombre = $nombreimagen;
         $limite_codigos = $_POST['limite_codigos'];
-        $actulizacion1 = "update revision_objeto set fecha_actualizacion = " . $fecha_actual . "  where id_revision_objeto = " . $id_revision_objeto . " AND id_empresa = " . $id_empresa . "";
+        $actulizacion1 = "update revision_objeto set fecha_actualizacion = " . $fecha_actual . "  where id_revision_objeto = " . $id_revision_objeto . " AND id_empresa = " . $_SESSION['idemp'] . "";
         if (!mysqli_query($pd, $actulizacion1)) {
             die('Error: ' . mysqli_error($pd));
         }
@@ -161,7 +148,7 @@ class obtener_cupon {
             die('Error: ' . mysqli_error($pd));
         }
         mysqli_close($pd);
-        header("Location:https://localhost/campeche-web2");
+        header("Location:https://localhost/campeche-web2/Controller/ControladorSitios.php");
     }
 
 }
